updating algorithm to ensure better solve
we changed the image library to PIL to have standard
RGB values and the edge matching algorithm to be
more robust - we normalized by the variance so that
edges with a low variance are de-prioritized, and we
added some constant factors to make sure that edges
that are a solid color are more de-prioritized. The
results seem improved on our sample images. We still
recommend the user to choose images with edges that
are not entirely white space.